requirements.txt
roboflow==1.1.16
ultralytics==8.1.3
numpy==1.26.3
matplotlib==3.8.2
ipython==8.20.0
opencv-python>=4.6.0
python-dotenv==1.0.0
keyboard==0.13.5
pytesseract==0.3.10
gTTS==2.5.0
paddlepaddle==2.6.0
paddleocr==2.7.0.3
google-generativeai==0.3.2
pygame==2.5.2
transformers==4.36.2
einops==0.7.0
timm==0.9.16
